Add docker-compose.yml with MySQL and Adminer services
This commit adds a docker-compose.yml file generating two services: MySQL and MariaDB 10.8.3 and Adminer. It sets up the root password for MySQL and allows users to uncomment a line for compatibility with Mac M1. This makes it easier to run and manage these local development dependencies in a Docker environment.

For adminer to work, you have to name the server mysql (default in the mariadb image?)