Refactor task view and update task listing logic
The task view has been refactored to include pagination links if there are any tasks. Also, the 'Add Task' link only shows when tasks exist. In addition to this, the task listing logic has been improved to showcase a paginated set of tasks instead of the previous logic that only fetched and displayed incompletes tasks.